requirements.txt
boto3
scikit-learn
pandas
numpy
python-dotenv
pydantic
fastapi
uvicorn